Commit 7

+++ src/main/java/com/auxdible/skrpg/player/economy/Bank.java
@@ -1,15 +1,15 @@
 package com.auxdible.skrpg.player.economy;
 
 public class Bank {
-    private int credits;
+    private double credits;
     private BankLevel level;
-    public Bank(int credits, BankLevel level) {
+    public Bank(double credits, BankLevel level) {
         this.credits = credits;
         this.level = level;
     }
-    public int getCredits() { return credits; }
+    public double getCredits() { return credits; }
     public BankLevel getLevel() { return level; }
     public void upgrade(BankLevel bankLevel) { this.level = bankLevel; }
-    public void deposit(int credits) { this.credits = this.credits + credits; }
-    public void withdraw(int credits) { this.credits = this.credits - credits; }
+    public void deposit(double credits) { this.credits = this.credits + credits; }
+    public void withdraw(double credits) { this.credits = this.credits - credits; }
 }
